Tests for pet package (#5)
* Added tests for rest.go in pet package. Tests for GET, POST and
DELETE endpoints.
Initial testing revealed an error in the Pet type, ID was an int,
should have been uint32. This has been corrected

* Finalized tests for rest related components of the pet package.
In particular, added testing for the PUT /api/pet/{key} endpoint.

* Added tests for in-memory storer implementation in the pet package.
Tests most common uses for CRUD operations, test quality depends
on reliability of sync.Map

* Updated README with testing instructions.
Updated .gitignore to ignore go test output files.

* fixup! test fixup

Replaced test variables with functions that return a new object
for each test. This avoids mutability from pollutng the tests.
Rewrote tests to follow given/when/then structure.

* fixup! test refactor

Refactored tests to work on the storer interface rather than
specific to the in-memory storage implementation.

* typo fix in README

* fixup! go mod tidy

* fixup!

Refactored rest_test to use setup rather than teardown.

* fixup!
Code cleanup for warnings by golangci-lint

+++ pkg/pet/memstore.go
@@ -22,13 +22,13 @@ func (m *MemStore) CreatePet(pet *Pet) error {
 	if _, ok := m.Load(pet.ID); ok {
 		return Errorf(ErrDuplicate, "Pet with id %d already exists", pet.ID)
 	}
-	m.Store(uint32(pet.ID), *pet)
+	m.Store(pet.ID, *pet)
 	return nil
 }
 
 // ReadPet gets a pet from the store given an ID
 func (m *MemStore) ReadPet(petID uint32) (*Pet, error) {
-	petData, ok := m.Load(petID)
+	petData, ok := m.Load(uint32(petID))
 	if !ok {
 		return nil, Errorf(ErrNotFound, "No pet exists with id %d", petID)
 	}
